refactor: Move suggestion matchers in a separate file
